+++ tests/phpunit/stubs/OCP/Files/GenericFileException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OCP\Files;
+
+if (!class_exists(GenericFileException::class)) {
+	class GenericFileException extends \Exception {
+	}
+}